Refactor page and add to index
